* fixed warnings

+++ src/client/cl_hunk.c
@@ -63,11 +63,11 @@ static void CL_ClientHunkDumpToFile_f (void)
 static void CL_ClientHunkUsage_f (void)
 {
 	ptrdiff_t size = clHunkPointerPos - clHunkData;
-	int allocated = (int)cl_hunkmegs->integer * 1024 * 1024;
+	int allocated = cl_hunkmegs->integer * 1024 * 1024;
 
 	Com_Printf("client hunk statistics:\n"
-		"used:      %10i bytes\n"
-		"free:      %10i bytes\n"
+		"used:      %10ti bytes\n"
+		"free:      %10ti bytes\n"
 		"allocated: %10i bytes\n", size, allocated - size, allocated);
 }
